% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfba.R
\docType{package}
\name{adfba}
\alias{adfba}
\alias{adfba-package}
\title{adfba: A package for Adaptive DFBA metabolic modeling.}
\description{
Implements the Adaptive DFBA technique proposed by Valverde et al. 2019.
}
\details{
The adfba Package provides the function
		adaptiveDFBA.
}
\section{adfba functions}{

\code{adaptiveDFBA} implements the Adaptive Dynamic Flux Balance Analysis
algorithm as described in Valverde et al. 2018.
}

\references{
Valverde, José R., Gullón, S., García-Herrero, C.A.,
Campoy, I, & Mellado, R.P. (2018) Dynamic modelling of 
overproduced secretory proteins in Streptomyces lividans 
using Adaptive DFBA. Submitted for publication.

Valverde, Jose R., Sonia Gullon, and Rafael P. Mellado.  (2018)
 	\emph{Modelling the metabolism of protein secretion through the Tat route in Streptomyces lividans}.
 	BMC microbiology 18.1: 59
}
\author{
\strong{Maintainer}: Jose R Valverde \email{jrvalverde@cnb.csic.es}

}
\keyword{internal}
