% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfba.R
\docType{data}
\name{Ec_core}
\alias{Ec_core}
\title{Escherichia coli Core Energy Metabolism Network}
\format{An object of class \code{modelorg}}
\source{
<http://gcrg.ucsd.edu/Downloads/EcoliCore>
}
\usage{
\code{data(Ec_core)}
}
\description{
The dataset is a network representation of the \emph{E. coli} core
 metabolism. It consists of 62 internal reactions, 14 exchange
 reactions and a biomass objective function.
}
\keyword{datasets}
