% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ecoli}
\alias{Ecoli}
\title{Escherichia coli Metabolic Network}
\format{An object of class \code{modelorg}}
\source{
<http://bigg.ucsd.edu/models/iJO1366>
}
\usage{
\code{
    data(Ecoli)
    lowbnd(Ecoli[react_id(Ec_core)=='EX_glc(e)']=-10;
    lowbnd(Ecoli)[react_id(Ec_core)=='EX_o2(e)']=-18;
    ## run adaptiveDFBA(), Ec_df will be an object of class optsol_dynamicFBA
    Ec_df <- adaptiveDFBA(Ecoli,
                          substrateRxns={'EX_glc(e)'},
                          initConcentrations=10,
                          initBiomass=.035,
                          timeStep=.25,
                          nSteps=17,
                          verboseMode=3)
    
    ## plot biomass and reactions
    plot(Ec_df,plotRxns=c('EX_glc(e)','EX_ac(e)'));
}
}
\description{
The dataset is a network representation of the \emph{E. coli} 
 metabolism as defined in the published metabolic model iJO1366.
 as published in
 
 Orth, J. D., Conrad, T. M., Na, J., Lerman, J. A., Nam, H., 
 Feist, A. M., & Palsson, B. O. (2011). A comprehensive 
 genome-scale reconstruction of Escherichia coli metabolism-2011. 
 Molecular systems biology, 7(1), 535.
 
 This model contains 1805 mtabolites, 2583 reactions and 1367 genes
}
\references{
Orth, J. D., Conrad, T. M., Na, J., Lerman, J. A., Nam, H., 
Feist, A. M., & Palsson, B. O. (2011). \emph{A comprehensive 
genome-scale reconstruction of Escherichia coli metabolism-2011}. 
Molecular systems biology, 7(1), 535.
}
\keyword{datasets}
