% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Slividans}
\alias{Slividans}
\title{Streptomyces lividans TK21, TK24 and 66 (1326) Metabolic Network}
\format{An object of class \code{modelorg}}
\source{
<https://bmcmicrobiol.biomedcentral.com/articles/10.1186/s12866-018-1199-3>
}
\usage{
\code{data(Slividans)}
}
\description{
The dataset is a network representation of the \emph{S. lividans} 
 metabolism as defined in the published metabolic model iJV1220.
 as published in
 
 Valverde, J. R., Gullón, S., & Mellado, R. P. (2018). Modelling the 
 metabolism of protein secretion through the Tat route in Streptomyces 
 lividans. BMC microbiology, 18(1), 59.
 
 It consists of over 1800 internal reactions, over 1200
 metabolites and the biomass objective function is called Biomass_SLI.
 You will need to customize it to your needs enabling/disabling the
 functions needed for the strain you want to model (compare with
 SlividansWT, Slividans_pIJ486, SlividansDAG, SlividansAML or SlividansTNF
 to see several examples).
}
\references{
Valverde, José R., Sonia Gullón, and Rafael P. Mellado. (2018)
 \emph{Modelling the metabolism of protein secretion through the Tat route 
 in Streptomyces lividans}. BMC microbiology 18.1: 59
}
\keyword{datasets}
