% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ec_core}
\alias{Ec_core}
\title{Escherichia coli Core Energy Metabolism Network}
\format{An object of class \code{modelorg}}
\source{
<http://gcrg.ucsd.edu/Downloads/EcoliCore>
}
\usage{
{
    data(Ec_core)
    lowbnd(Ec_core)[react_id(Ec_core)=='EX_glc(e)']=-10;
    lowbnd(Ec_core)[react_id(Ec_core)=='EX_o2(e)']=-18;
    ## run adaptiveDFBA(), Ec_df will be an object of class optsol_dynamicFBA
    Ec_df <- adaptiveDFBA(Ec_core,
                          substrateRxns={'EX_glc(e)'},
                          initConcentrations=10,
                          initBiomass=.035,
                          timeStep=.25,
                          nSteps=17,
                          verboseMode=3)
    
    ## plot biomass and reactions
    plot(Ec_df,plotRxns=c('EX_glc(e)','EX_ac(e)'));
}
}
\description{
The dataset is a network representation of the \emph{E. coli} core
 metabolism. It consists of 62 internal reactions, 14 exchange
 reactions and a biomass objective function.
}
\keyword{datasets}
